Move ServiceManager and Service classes to engine assembly
