+++ src/TestEngine/testcentric.engine.core/testcentric.engine.core.csproj
@@ -35,4 +35,7 @@
     <PackageReference Include="NUnit.Engine.Api" Version="3.11.1" />
     <PackageReference Include="TestCentric.Metadata" Version="1.7.0" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Services\" />
+  </ItemGroup>
 </Project>
